Move client id tests under dedicated testing class
